+++ custom-recipes/neo4j-export-nodes/recipe.py
@@ -1,7 +1,7 @@
 import os
 from dataiku.customrecipe import get_recipe_config
 from commons import get_input_output, create_dataframe_iterator, check_load_from_csv, ImportFileHandler
-from dku_neo4j import NodesExportParams, Neo4jHandle
+from dku_neo4j.neo4j_handle import NodesExportParams, Neo4jHandle
 
 # --- Setup recipe
 recipe_config = get_recipe_config()
